AgentMCP/requirements.txt
google-adk>=0.3.0
mcp>=1.0.0
httpx>=0.27.0
sqlalchemy>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
RestrictedPython>=7.0
uvicorn>=0.30.0
aiosqlite>=0.20.0
